Refactored details page and side_strike_text to match naming scheme
