+++ lib/widgets/side_strike_text.dart
